requirements.txt
asgiref==3.6.0
autopep8==2.0.1
certifi==2019.3.9
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.2.0
defusedxml==0.6.0
distlib==0.3.6
Django==4.1.4
django-allauth==0.39.1
django-autoslug==1.9.9
django-countries==5.3.3
django-crispy-forms==1.7.2
django-debug-toolbar==1.10.1
django-taggit==4.0.0
django-tinymce==3.6.1
filelock==3.8.2
idna==2.8
libpng-bins==0.0.3
mysqlclient==2.1.1
numpy==1.25.2
oauthlib==3.0.1
pep8==1.7.1
Pillow==9.3.0
platformdirs==2.6.0
pycodestyle==2.10.0
pycparser==2.21
pylibjpeg-libjpeg==1.3.4
python-decouple==3.1
python3-openid==3.1.0
pytz==2018.5
requests==2.21.0
requests-oauthlib==1.2.0
sqlparse==0.4.3
stripe==2.27.0
tzdata==2022.7
urllib3==1.24.2
virtualenv==20.17.1
virtualenvwrapper-win==1.2.7
